Create a base tasks builder
